+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Feature
 
 - Enabling support for the new active LTS NodeJS version 14. NodeJS 10 eol will happen on 2021-04-30 and Volto will update accordingly. More information on https://nodejs.org/en/about/releases @sneridagh
+- Add `settings.backendResourceMatch` to allow handling of backend resources similar to `@@images` or `@@download` @tiberiuichim
 
 ### Bugfix